After meeting with Guy
First change is to the way that the inverse Fourier transform is computed to calculate the time-series.

Changed kernel to SHO term, but issues with having just that and for some reason keeps wanting a real term kernel too.

Needs a good combing through the celerite section to ensure it's working properly before setting up to work with multiple Lorentz peaks