Working GP_test (Jitter & SHO kernels / 1 peak)
The GP_test works for one peak with the data in units of days, not seconds; hence the frequency values are less intuitive to interpret.

Contains a jitter term kernel (to handle the offset and white noise), and an SHO term kernel to give the properties of the oscillations.

The second GP_test(2) notebook was added to do this same thin, but with Hz units.

Now it is necessary to build on this and simulate more peaks. In addition, it's necessary to determine whether another kernel must be added in order to compute a background level (as it's set to one currently and handled by the jitter term).