Working for daily SMMF / created new for 40s cadence
Working still for daily SMMF.

Trying to create a new file which has 40s cadence data for 3 months, in order to test on something more representative of the  month SMMF chunks.

Bill suggested that perhaps the kernel will struggle with the lifetime due to the resolution, but could do all TS modelling (even for background) with the GP kernels.

Unsure if celerite can handle zero-centred Lorentzian for ER background...