40s cadence test
Run for 27 day peak and fit produces a 28.4 (+1.7/-2.0) days. The FT of the IFT (i.e. TS) gives a peak @ 0.3858uHz, so approx 30 days. Therefore this has worked to some extent.

The corner plot does not looks great... but burn-in was 5000 and production was 15000.

Don't think a longer burn-in would necessarily help here, and running out of the ideas for this working... could perhaps have a look at the DFM paper asteroseismic example instead(?).